app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
redis==3.5.3
psycopg2-binary==2.8.6
